+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# REPLACE_GLOBALLY_WITH_NEXT_VERSION (3.2.2)
+- PPI-219 - Disable PayPal on Carts with total price of 0
+
 # 3.2.1
 - PPI-279, PPI-297 - Extends the partner referral API
 - PPI-290 - Improved extensibility